# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Yunora AI Question Generation Admin Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /standards:
    get:
      operationId: listStandards
      tags: [standards]
      summary: List standards
      security:
        - bearerAuth: []
      parameters:
        - name: boardId
          in: query
          schema: { type: integer }
        - name: search
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: List of standards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardList"
    post:
      operationId: createStandard
      tags: [standards]
      summary: Create a standard
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StandardInput"
      responses:
        "201":
          description: Created standard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Standard"
  /standards/{id}:
    patch:
      operationId: updateStandard
      tags: [standards]
      summary: Update standard
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StandardUpdate"
      responses:
        "200":
          description: Updated standard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Standard"
    delete:
      operationId: deleteStandard
      tags: [standards]
      summary: Delete standard
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── SUBJECTS ─────────────────────────────────────────────────────────────────
  /subjects:
    get:
      operationId: listSubjects
      tags: [subjects]
      summary: List subjects
      security:
        - bearerAuth: []
      parameters:
        - name: standardId
          in: query
          schema: { type: integer }
        - name: boardId
          in: query
          schema: { type: integer }
        - name: search
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubjectList"
    post:
      operationId: createSubject
      tags: [subjects]
      summary: Create a subject
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
  /subjects/{id}:
    patch:
      operationId: updateSubject
      tags: [subjects]
      summary: Update subject
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectUpdate"
      responses:
        "200":
          description: Updated subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
    delete:
      operationId: deleteSubject
      tags: [subjects]
      summary: Delete subject
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── CHAPTERS ─────────────────────────────────────────────────────────────────
...
  /questions:
    get:
      operationId: listQuestions
      tags: [questions]
      summary: List questions with filters
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: boardId
          in: query
          schema: { type: integer }
        - name: standardId
          in: query
          schema: { type: integer }
        - name: subjectId
          in: query
          schema: { type: integer }
        - name: chapterId
          in: query
          schema: { type: integer }
        - name: topicId
          in: query
          schema: { type: integer }
        - name: difficulty
          in: query
          schema: { type: string, enum: [easy, medium, hard, advanced] }
        - name: questionType
          in: query
          schema: { type: string }
        - name: model
          in: query
          schema: { type: string }
        - name: dateFrom
          in: query
          schema: { type: string, format: date }
        - name: dateTo
          in: query
          schema: { type: string, format: date }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Paginated questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionList"
  /questions/{id}:
    get:
      operationId: getQuestion
      tags: [questions]
      summary: Get a single question
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
    patch:
      operationId: updateQuestion
      tags: [questions]
      summary: Edit a question
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionUpdate"
      responses:
        "200":
          description: Updated question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
    delete:
      operationId: deleteQuestion
      tags: [questions]
      summary: Delete a question
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── AI GENERATION ─────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Standard:
      type: object
      required: [id, name, boardId]
      properties:
        id: { type: integer }
        name: { type: string }
        level: { type: integer }
        boardId: { type: integer }
        boardName: { type: ["string", "null"] }
        isActive: { type: boolean }
        subjectsCount: { type: integer }
        createdAt: { type: string, format: date-time }
    StandardInput:
      type: object
      required: [name, boardId, level]
      properties:
        name: { type: string }
        level: { type: integer }
        boardId: { type: integer }
        isActive: { type: boolean, default: true }
    StandardUpdate:
      type: object
      properties:
        name: { type: string }
        level: { type: integer }
        isActive: { type: boolean }
    StandardList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Standard"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }

    # ── SUBJECTS ──────────────────────────────────────────────────────────────
    Subject:
      type: object
      required: [id, name, standardId]
      properties:
        id: { type: integer }
        name: { type: string }
        code: { type: ["string", "null"] }
        standardId: { type: integer }
        standardName: { type: ["string", "null"] }
        boardId: { type: integer }
        boardName: { type: ["string", "null"] }
        isActive: { type: boolean }
        chaptersCount: { type: integer }
        questionsCount: { type: integer }
        createdAt: { type: string, format: date-time }
    SubjectInput:
      type: object
      required: [name, standardId]
      properties:
        name: { type: string }
        code: { type: string }
        standardId: { type: integer }
        isActive: { type: boolean, default: true }
    SubjectUpdate:
      type: object
      properties:
        name: { type: string }
        code: { type: string }
        isActive: { type: boolean }
    SubjectList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Subject"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }

    # ── CHAPTERS ──────────────────────────────────────────────────────────────
...
    Question:
      type: object
      required: [id, question, questionType, difficulty, difficultyScore, correctAnswer, explanation, qualityScore]
      properties:
        id: { type: integer }
        question: { type: string }
        questionType: { type: string }
        difficulty: { type: string, enum: [easy, medium, hard, advanced] }
        difficultyScore: { type: integer }
        correctAnswer: { type: string }
        options: { type: ["string", "null"] }
        explanation: { type: string }
        qualityScore: { type: number }
        estimatedSolveTime: { type: integer }
        learningObjective: { type: ["string", "null"] }
        topicId: { type: ["integer", "null"] }
        topicName: { type: ["string", "null"] }
        chapterId: { type: ["integer", "null"] }
        chapterName: { type: ["string", "null"] }
        subjectId: { type: ["integer", "null"] }
        subjectName: { type: ["string", "null"] }
        boardId: { type: ["integer", "null"] }
        boardName: { type: ["string", "null"] }
        standardId: { type: ["integer", "null"] }
        standardName: { type: ["string", "null"] }
        modelUsed: { type: ["string", "null"] }
        providerId: { type: ["integer", "null"] }
        generatedAt: { type: string, format: date-time }
    QuestionUpdate:
      type: object
      properties:
        question: { type: string }
        correctAnswer: { type: string }
        options: { type: string }
        explanation: { type: string }
        difficulty: { type: string, enum: [easy, medium, hard, advanced] }
        difficultyScore: { type: integer }
        learningObjective: { type: string }
    QuestionList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }

    # ── GENERATION ────────────────────────────────────────────────────────────
...